+++ Engine/MainThread/UStaticMesh.h
@@ -20,7 +20,7 @@ public:
         return mVertexes;
     }
 
-    inline const std::vector<uint16>& GetIndexes() const
+    inline const std::vector<uint32>& GetIndexes() const
     {
         return mIndexes;
     }
@@ -32,6 +32,6 @@ public:
 
 private:
     std::vector<FStaticMeshVertex> mVertexes;
-    std::vector<uint16> mIndexes;
+    std::vector<uint32> mIndexes;
     FRHIVertexLayout mVertexLayout;
 };
